# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MagenticOne Crypto Analysis API
  description: |
    Multi-agent cryptocurrency analysis platform API.
    
    This API provides:
    - Real-time chat with AI agents for crypto analysis
    - Chart generation with technical indicators
    - Secure exchange credential management
    - WebSocket streaming for live agent updates
  version: 1.0.0
  contact:
    name: MagenticOne Team
  license:
    name: MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: http://backend:8000/api/v1
    description: Docker internal
...
paths:
...
  /settings/exchange:
    post:
      tags: [Settings]
      summary: Save exchange credentials
      description: |
        Save encrypted exchange credentials (Bitget).
        Credentials are encrypted with AES-256 before storage.
      operationId: saveExchangeCredentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExchangeCredentials'
      responses:
        '200':
          description: Credentials saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Invalid credentials

    get:
      tags: [Settings]
      summary: Get exchange status
      description: Check if exchange credentials are configured (does not return secrets)
      operationId: getExchangeStatus
      responses:
        '200':
          description: Exchange status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExchangeStatus'

    delete:
      tags: [Settings]
      summary: Remove exchange credentials
      description: Delete stored exchange credentials
      operationId: deleteExchangeCredentials
      responses:
        '200':
          description: Credentials removed
  /settings/llm:
    post:
      tags: [Settings]
      summary: Configure LLM provider
      description: Set the LLM provider configuration
      operationId: configureLLM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LLMConfig'
      responses:
        '200':
          description: LLM configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

    get:
      tags: [Settings]
      summary: Get LLM configuration
      description: Get current LLM provider settings
      operationId: getLLMConfig
      responses:
        '200':
          description: Current LLM config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LLMConfigResponse'

# =============================================================================
# Components
# =============================================================================
components:
  schemas:
...
    ExchangeCredentials:
      type: object
      required: [api_key, api_secret, passphrase]
      properties:
        api_key:
          type: string
          minLength: 1
        api_secret:
          type: string
          minLength: 1
        passphrase:
          type: string
          minLength: 1
    ExchangeStatus:
      type: object
      properties:
        configured:
          type: boolean
        exchange:
          type: string
          example: bitget
        connected:
          type: boolean
        last_checked:
          type: string
          format: date-time
    LLMConfig:
      type: object
      required: [provider]
      properties:
        provider:
          type: string
          enum: [azure]
        model:
          type: string
          description: Model/deployment name
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 0.7
    LLMConfigResponse:
      type: object
      properties:
        provider:
          type: string
        model:
          type: string
        connected:
          type: boolean

    # Common Schemas
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, warning, error]
        message:
          type: string
        details:
          type: object
...